requirement.txt
pandas 2.2.2
plotly_express 5.24.1
streamlit 1.37.1
